refactored replacement selection sort

+++ include/bucket_manager.h
@@ -8,7 +8,7 @@
 class BucketManager {
 public:
   BucketManager(Collection& collection, std::fstream& buffer,
-                std::vector<int> buckets_sizes);
+                std::vector<size_t> buckets_sizes);
   void init();
   Bucket pop();
   void heapify(std::vector<std::string> keys);
@@ -19,11 +19,11 @@ public:
  private:
   Collection& collection;
   std::vector<Bucket> buckets_;
-  std::vector<int> bucket_sizes_;
+  std::vector<size_t> bucket_sizes_;
   std::vector<size_t> offsets_;
   std::vector<size_t> cursors;
   std::fstream& sorted_buffer;
-  std::vector<size_t> find_offsets(const std::vector<int>& bucket_sizes);
+  std::vector<size_t> find_offsets(const std::vector<size_t>& bucket_sizes);
   Bucket fetch(size_t bucket_number);
 };
